Fix NPE when re-synch project in a large workspace.
When re-sync with version control or updating dependencies directly on a
project, there is a higher change that the type for java.lang.Object
fails due to change on the JDT projects nameEnviroment.

This is a quick-fix until we find a way to see if the analyzer job can
be delayed.